Add vertex color toggle to standard material and add tooltips

+++ Shaders/ConfigurableCutoutUnlit.shader
@@ -22,14 +22,14 @@ Shader "Configurable/Unlit/Cutout"
 		_Cutoff ("Alpha cutoff", Range(0,1)) = 0.5
 		
 		[Header(Rendering)]
-		_Offset("Offset", float) = 0
+		_Offset("Offset", Float) = 0
 		[Enum(UnityEngine.Rendering.CullMode)] _Culling ("Cull Mode", Int) = 2
 		[Enum(Off,0,On,1)] _ZWrite("ZWrite", Int) = 1
 		[Enum(UnityEngine.Rendering.CompareFunction)] _ZTest ("ZTest", Int) = 4
 		[Enum(None,0,Alpha,1,Red,8,Green,4,Blue,2,RGB,14,RGBA,15)] _ColorMask("Color Mask", Int) = 15
 		
 		[Header(Stencil)]
-		_Stencil ("Stencil ID [0;255]", Float) = 0
+		_Stencil ("Stencil ID [0;255]", Int) = 0
 		[Enum(UnityEngine.Rendering.CompareFunction)] _StencilComp ("Stencil Comparison", Int) = 0
 		[Enum(UnityEngine.Rendering.StencilOp)] _StencilOp ("Stencil Operation", Int) = 0
 		[Enum(UnityEngine.Rendering.StencilOp)] _StencilFail ("Stencil Fail", Int) = 0
